Desarrollo creado la funcionalidad de evento y creado el front de enlace publico.

+++ src/main/java/com/tufg/gestor_reuniones/repository/DisponibilidadRepository.java
@@ -3,5 +3,5 @@ package com.tufg.gestor_reuniones.repository;
 import com.tufg.gestor_reuniones.model.Disponibilidad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DisponibilidadRepository extends JpaRepository<Disponibilidad, Long> {
+public interface DisponibilidadRepository extends JpaRepository<Disponibilidad, Integer> {
 }
